requirements.txt
streamlit>=1.30.0
python-dotenv==1.0.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.2
joblib>=1.2.0
ollama
requests>=2.31.0
pymongo>=4.5.0
